refactor(vault): move Filter class to :uimodels
