+++ core/uimodel/src/commonMain/composeResources/values/strings.xml
@@ -1,29 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <!--Auth-->
-    <string name="search">Search</string>
-    <string name="vault">Vault</string>
     <string name="pass">Pass</string>
-    <string name="guard">Guard</string>
-    <string name="buttons">Buttons</string>
-    <string name="privacy_policy_link">https://andrewdobosh.notion.site/Privacy-Policy-3f95b4857e0e49a78b35f0ca837b4ec5</string>
-    <string name="terms_of_use_link">https://andrewdobosh.notion.site/Terms-of-use-bb7607b71dbc40d9a336c9ed38d5a7e8</string>
-    <string name="accept_and_agree">I accept and agree to the </string>
-    <string name="and"> and </string>
-    <string name="terms_of_use"> Terms of Use</string>
-    <string name="privacy_policy">Privacy Policy.</string>
-
-    <string name="password_error_bad_length">Password must contain at least 8 characters.</string>
-    <string name="password_error_no_numbers">Password must contain at least one number.</string>
-    <string name="password_error_no_capital">Password must contain at least one uppercase letter.</string>
-    <string name="password_error_no_small">Password must contain at least one lowercase letter.</string>
-    <string name="email_error_incorrect">Email is no valid</string>
-
-    <string name="tools">Tools</string>
-    <string name="settings">Settings</string>
-
-
-
-    <!--Content Description-->
-    <string name="cd_app_logo">App Logo</string>
 </resources>
