+++ feature/vault/domain/build.gradle.kts
@@ -35,6 +35,7 @@ kotlin {
             api(project(":core:model"))
             api(project(":core:common"))
             api(project(":core:ui"))
+            api(project(":core:uimodel"))
             implementation(project(":feature:auth:data"))
 
             implementation(libs.ktor.serialization.kotlinx.json)
