+++ feature/vault/domain/src/commonTest/kotlin/com/thejohnsondev/domain/ItemTypeFilterChangeUseCaseImplTest.kt
@@ -1,6 +1,6 @@
 package com.thejohnsondev.domain
 
-import com.thejohnsondev.ui.model.Filter
+import com.thejohnsondev.uimodel.Filter
 import kotlin.test.Test
 import kotlin.test.assertEquals
