+++ feature/vault/presentation/build.gradle.kts
@@ -42,6 +42,7 @@ kotlin {
             api(project(":core:model"))
             api(project(":core:common"))
             api(project(":core:ui"))
+            api(project(":core:uimodel"))
             api(project(":feature:vault:domain"))
 
             implementation(libs.ktor.serialization.kotlinx.json)
